requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-generativeai==0.8.5
python-dotenv==1.1.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.3
httpx==0.28.1
